requirements.txt
streamlit
streamlit-autorefresh
pandas
numpy
requests
python-dotenv
yfinance
transformers
torch
sentencepiece
accelerate
keybert
spacy
nltk
plotly
feedparser
